change memory allocator and check return values
Before this commit we used malloc and free on pointers allocated by rcutils_get_default_allocator.
They should be the same, but just to be sure for all platforms.
We should also check the return values of methods that fail (allocator and string assignment).
Given that much changes I wanted to make I copied several functions from rosidl_runtime_c and adjusted them for the needs here and replaced the old ones.